build(esm): ESM main output + renderer code-split for streamdown/katex
Flip the electron-vite main bundle to ESM (out/main/index.mjs) and split the
~2 MB streamdown/katex/@streamdown/math markdown renderer into a lazy chunk.

Main process:
- electron.vite.config.ts: main.build.rollupOptions.output.format 'es'
- package.json: main -> out/main/index.mjs
- __dirname -> dirname(fileURLToPath(import.meta.url)) in main/index.ts and
  services/ocr/rasterizer.ts (test-suite idiom; db.ts createRequire unchanged)
- Preload STAYS CommonJS: Electron only allows ESM preload with sandbox:false,
  which both windows never do (spec §0). So format:'es' is scoped to main and
  "type":"module" is deliberately not set.

Bundle:
- package.json "sideEffects":["*.css"] so Rollup drops unused own-module exports
- AssistantMarkdown (used on chat/translate/documents/images) moved to its own
  module behind a React.lazy wrapper (AssistantMarkdownLazy), re-exported as
  AssistantMarkdown from the chat barrel -> loads as a separate async chunk on
  first answer render. Initial index-*.js: 3,196 kB -> 1,194 kB (-63%).
- vitest.config.ts aliases ./AssistantMarkdownLazy -> the real component so
  render tests don't await Suspense; two direct test importers repointed.

typecheck clean; build green (main is ESM, preload CJS); suite 3812/44 (the lone
fail is the pre-existing env-only vision-status test, fails on the pristine tree).

Docs: packaging.md "Module format & renderer bundle — design record";
architecture.md process-model preload/main lines; BUILD_STATE.md entry.
