+++ BUILD_STATE.md
@@ -6,6 +6,9 @@
 > It carries: current status, decisions, shared data contracts, next actions, open issues.
 
 
+_2026-07-07 — **ESM main output + renderer code-split — SHIPPED on branch `esm`, UNPUSHED. NOT committed until owner OKs.**_
+_A build-tooling change, no product behaviour change. **(1) Main process now bundles as ESM** (`out/main/index.mjs`): `electron.vite.config.ts` sets `main.build.rollupOptions.output.format:'es'`; `apps/desktop/package.json` `main` → the `.mjs`. `__dirname` (gone in ESM) reconstructed via `dirname(fileURLToPath(import.meta.url))` in `src/main/index.ts` + `services/ocr/rasterizer.ts` (the test-suite idiom); `db.ts`'s `createRequire` for `node:sqlite` is ESM-native, untouched. **Preload STAYS CommonJS** (`out/preload/index.js`,`ocr.js`) — Electron only allows ESM preload with `sandbox:false`, which both windows never do (spec §0); so `format:'es'` is scoped to `main` only and `"type":"module"` is deliberately NOT set (it would force `.js`→ESM and break the CJS preload). Main→preload refs still point at `../preload/index.js`. A "pure ESM" app is unreachable while sandboxed; main-ESM + preload-CJS is the end state. **(2) `"sideEffects":["*.css"]`** added so Rollup can drop unused exports from our own modules (tree-shaking keys off the ESM *input* graph, already ESM — the ESM-output flip itself buys none). **(3) Renderer code-split — the real bundle win.** streamdown+katex+@streamdown/math (~2 MB) were eager in the initial chunk because `AssistantMarkdown` (used on chat/translate/documents-preview/images) statically imported them. Moved the Streamdown/KaTeX renderer into its own module `src/renderer/chat/AssistantMarkdown.tsx` behind a `React.lazy` wrapper `AssistantMarkdownLazy.tsx` (re-exported as `AssistantMarkdown` from the chat barrel), so it loads as a **separate async chunk on first answer render**. Initial `index-*.js`: **3,196 kB → 1,194 kB (−63%)**. The async chunk is Vite-named `mermaid-*` (a streamdown-internal module) but holds only streamdown+katex — verified NO real mermaid/cytoscape/d3 (the asar mermaid exclusion in packaging.md still holds). Suspense fallback renders raw text in a `.md` box (local, few-ms load, no flash). **Tests:** two direct importers repointed `Transcript`→`./AssistantMarkdown`; `vitest.config.ts` aliases `./AssistantMarkdownLazy`→the real component so render assertions don't await Suspense. **Verify:** `typecheck` clean; `npm run build` green (confirmed `index.mjs` is ESM, preload is CJS `"use strict";require(...)`); **suite 3812/44** — the lone fail is the pre-existing env-only `vision-status` test (a real vision runtime on this dev box leaks past its stub; fails identically on the pristine tree). **Docs:** `packaging.md` gained "Module format & renderer bundle — design record"; `architecture.md` Process-model preload/main lines note the CJS/ESM split. **Next actions: owner review → commit on `esm` (nothing committed yet)**; a `package`-then-smoke of the produced `.exe` is the remaining manual gate (electron-builder must pick up `out/main/index.mjs`)._
+
 _2026-07-07 — **Issue #31 (translation view) — BOTH FIXES SHIPPED + RELEASE v0.1.45 — on local `master`, UNPUSHED.**_
 _**(a) False "couldn't finish" banner on every SUCCESSFUL translation — fixed.** Root cause: `isCleanStop` (TA-5 M6) keyed on the LEGACY llama-server final-frame fields (`stopping_word`/`stopped_eos`); the pinned b9849 reports the consolidated **`stop_type`** field (`eos|word|limit|none`) instead, and Gemma ends a finished turn on `<end_of_turn>` as an EOS-class token ⇒ the real success frame is `stop_type:"eos"` with an EMPTY `stopping_word` ⇒ every clean window was misread as a limit stop → the view job failed `runtimeFailed` AFTER the full text had streamed (the issue's always-on banner); the doc-task path likewise burned a duplicate ~30-min retry decode per window then marked it failed. Fix in `services/translation/completion.ts`: `CompletionFinal.stopType` parsed from the frame; `isCleanStop` = `stop_type ∈ {eos,word}` OR the legacy signals (older shapes + scripted test translators stay valid); a REAL `stop_type:"limit"` still fails the window. Regression-pinned with the exact b9849 success/limit frames (`translation-completion.test.ts`); the manual smoke's window log now prints `stop=<stop_type>`. **(b) Language set widened (owner decision on the issue): curated 10 → the 51-code WMT24++ production tier** (TranslateGemma's 55 evaluated locales collapsed to bare codes; `zh`=Simplified; the ~105 experimental template languages stay OUT). Still a CLOSED, server-validated list — `TRANSLATION_LANGUAGE_CODES` + native/English name maps widened (`shared/types.ts` + `translation/prompt.ts`); pickers render all 51 (native names), defaults stay de/en; planner stays safe on the new space-less scripts (`approxTokenCount` charges them per-character ⇒ over-chunk only, never overflow). The smoke keeps its curated-10 calibration samples (new `SMOKE_LANGS`); the `'ja'`-as-invalid doc-task test fixture retargeted to `'xy'`. **Docs:** architecture.md "Translation sidecar" record gained the issue-#31 bullet (+ D5 row superseded-in-extent note); known-limitations language bullet rewritten (51, evidence split); model-benchmarks §11.2 widening note. **Release:** version bumped 0.1.44 → **0.1.45**, tag **v0.1.45**. **Next actions: owner pushes `master` + tags** (`git push origin master --follow-tags`) — pushing closes issue-#31 follow-up; the smoke's V1 template-name reconcile (fa/zh spellings) rides the next manual smoke run._
